+++ piquasso/_simulators/gaussian/probabilities.py
@@ -28,6 +28,8 @@ from piquasso.api.connector import BaseConnector
 
 class DensityMatrixCalculation(abc.ABC):
     _normalization: float
+    _A: np.ndarray
+    _b: np.ndarray
 
     def __init__(self, connector: BaseConnector):
         self.connector = connector
@@ -46,15 +48,9 @@ class DensityMatrixCalculation(abc.ABC):
         )
 
     def get_density_matrix(self, occupation_numbers: np.ndarray) -> np.ndarray:
-        n = occupation_numbers.shape[0]
-
-        density_matrix = np.empty(shape=(n, n), dtype=complex)
-
-        for i, bra in enumerate(occupation_numbers):
-            for j, ket in enumerate(occupation_numbers):
-                density_matrix[i, j] = self.get_density_matrix_element(bra, ket)
-
-        return density_matrix
+        return self.connector.density_matrix_from_representation(
+            self._A, self._b, self._normalization, occupation_numbers
+        )
 
     def get_particle_number_detection_probabilities(
         self, occupation_numbers: np.ndarray
@@ -103,6 +99,8 @@ class NondisplacedDensityMatrixCalculation(DensityMatrixCalculation):
 
         self._A = X @ (np.identity(2 * d, dtype=complex) - Qinv)
 
+        self._b = np.zeros(2 * d, dtype=complex)
+
         self._normalization: float = 1 / np.sqrt(np.linalg.det(Q))
 
     def calculate_hafnian(self, reduce_on: np.ndarray) -> float:
@@ -137,6 +135,8 @@ class DisplacedDensityMatrixCalculation(DensityMatrixCalculation):
         self._A = X @ (np.identity(2 * d, dtype=complex) - Qinv)
 
         self._gamma = complex_displacement.conj() @ Qinv
+
+        self._b = self._gamma
 
         self._normalization: float = np.exp(
             -0.5 * self._gamma @ complex_displacement